Fixed powershell script bug

+++ scripts/baseshaders.ps1
@@ -1,6 +1,10 @@
 $bro=$args[0]
 $bro2=$args[1]
-rm -r $bro2/BaseShaders
+
+if (Test-Path -Path $bro2/BaseShaders) {
+  rm -r $bro2/BaseShaders
+}
+
 cp -r $bro/Application/Renderer/BaseShaders -Destination $bro2
 cd $bro2/BaseShaders
 $files = Get-ChildItem $bro2/BaseShaders
@@ -9,9 +13,9 @@ foreach($file in $files)
 {
     $extn = [IO.Path]::GetExtension($file)
     if ($extn -eq ".glsl" ){
-    $baseName = (Get-Item $file).Basename
-    $spvName = $baseName + ".spv"
-    glslangValidator -V $file -o  $spvName
+      $baseName = (Get-Item $file).Basename
+      $spvName = $baseName + ".spv"
+      glslangValidator -V $file -o  $spvName
     }
 }
